Add CI system
Test on all operating system and python version combinations

Don't test the notebooks; just see if we can get all dependencies to
install

The challenge A notebook requires a token, and the challenge B
notebook takes more than 5 minutes to run a cell on my notebook.
So, as a first step, let's just use CI to ensure that all dependencies
can be installed on all platforms successfully.

Make the tox environment do nothing

This way, it at least installs the deps

Drop python 3.9 as a supported version

```
INFO: pip is looking at multiple versions of qiskit-ibm-catalog to determine which version is compatible with other requirements. This could take a while.

The conflict is caused by:
    The user requested qiskit_serverless
    qiskit-ibm-catalog 0.0.0 depends on qiskit-serverless>=0.14.2
```